test(components-react): add carousel e2e test verifying lifecycles | bh | #2382
