+++ packages/components-react/tests/e2e/specs/stencil-lifecycles.e2e.ts
@@ -1,5 +1,5 @@
 import type { Page } from 'puppeteer';
-import { getLifecycleStatus, goto, selectNode, waitForComponentsReady } from '../helpers';
+import { getLifecycleStatus, goto, selectNode, trackLifecycleStatus, waitForComponentsReady } from '../helpers';
 
 let page: Page;
 beforeEach(async () => (page = await browser.newPage()));
@@ -11,40 +11,7 @@ it('should not trigger updates on non-default props', async () => {
   await client.send('Emulation.setCPUThrottlingRate', { rate: 6 });
 
   await goto(page, 'stencil-lifecycles');
-
-  await page.evaluate(() => {
-    const script = document.createElement('script');
-    script.text = `
-      const LIFECYCLE_STATUS_KEY = 'stencilLifecycleStatus';
-
-      // initial status
-      window[LIFECYCLE_STATUS_KEY] = {
-        componentWillLoad: { all: 0 },
-        componentDidLoad: { all: 0 },
-        componentWillUpdate: { all: 0 },
-        componentDidUpdate: { all: 0 },
-      };
-
-      const hooks = ['componentWillLoad', 'componentDidLoad', 'componentWillUpdate', 'componentDidUpdate'];
-      for (let hook of hooks) {
-        window.addEventListener(\`stencil_\${hook}\`, (e) => {
-          const eventName = e.type.replace('stencil_', '');
-          const tagName = e.composedPath()[0].tagName.toLowerCase();
-
-          if (window[LIFECYCLE_STATUS_KEY][eventName][tagName] === undefined) {
-            // to ensure the lifecycle hook is not undefined in our e2e test, we have to initialize it
-            for (const hook of hooks) {
-              window[LIFECYCLE_STATUS_KEY][hook][tagName] = 0;
-            }
-          }
-
-          window[LIFECYCLE_STATUS_KEY][eventName][tagName]++;
-          window[LIFECYCLE_STATUS_KEY][eventName].all++;
-
-        });
-      }`;
-    document.body.appendChild(script);
-  });
+  await trackLifecycleStatus(page);
 
   const button = await selectNode(page, 'button');
   const status = await getLifecycleStatus(page);
